Moved DiagonalBlindsBG into Backgrounds package.
